backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
motor==2.5.1
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
pymongo==3.12.3
dnspython==2.5.0